refactor(kernels): extract opencl work-size optimizer microcrate
