+++ crates/bitnet-kernels/Cargo.toml
@@ -19,10 +19,8 @@ include = [
 
 [dependencies]
 bitnet-common = { path = "../bitnet-common", version = "0.2.1-dev" }
-bitnet-cpu-activations = { path = "../bitnet-cpu-activations", version = "0.2.1-dev" }
-bitnet-cpu-detect = { path = "../bitnet-cpu-detect", version = "0.2.1-dev", default-features = false }
 bitnet-device-probe = { path = "../bitnet-device-probe", version = "0.2.1-dev" }
-bitnet-simd = { path = "../bitnet-simd", version = "0.2.1-dev" }
+bitnet-opencl-work-size = { path = "../bitnet-opencl-work-size", version = "0.2.1-dev" }
 thiserror.workspace = true
 bytemuck.workspace = true
 rayon.workspace = true
@@ -32,6 +30,7 @@ opencl3 = { version = "0.9", optional = true }
 cc = { version = "1.2.46", optional = true }
 bindgen = { version = "0.72.1", optional = true }
 half = { version = "2.7.1", optional = true }
+libm = "0.2.16"
 sysinfo = "0.37.2"
 memory-stats = "1.2.0"
 
@@ -58,10 +57,6 @@ rand_chacha = "0.9.0"
 bitnet-tests = { path = "../../tests", version = "0.2.1-dev" }
 bitnet-test-support.workspace = true
 
-[target.'cfg(all(target_os = "macos", target_arch = "aarch64"))'.dev-dependencies]
-wgpu = { version = "24", features = ["metal"] }
-pollster = "0.4"
-
 [target.'cfg(target_arch = "x86_64")'.dependencies]
 # x86-specific dependencies
 
@@ -75,9 +70,9 @@ npu-backend = []
 cpu = ["cpu-fallback"]
 cpu-fallback = []
 cpu-optimized = ["avx512", "avx2", "neon"]
-avx2 = ["bitnet-cpu-detect/avx2"]
-avx512 = ["bitnet-cpu-detect/avx512"]
-neon = ["bitnet-cpu-detect/neon"]
+avx2 = []
+avx512 = []
+neon = []
 cuda = ["dep:cudarc", "dep:half", "bitnet-device-probe/cuda"]
 rocm = ["bitnet-device-probe/rocm"]
 hip = ["rocm"]
@@ -97,18 +92,6 @@ harness = false
 [[test]]
 name = "kernel_tests"
 path = "tests/kernel_tests.rs"
-
-[[test]]
-name = "apple_silicon_benchmarks"
-path = "tests/apple_silicon_benchmarks.rs"
-
-[[test]]
-name = "neon_correctness_exhaustive"
-path = "tests/neon_correctness_exhaustive.rs"
-
-[[test]]
-name = "metal_device_integration_tests"
-path = "tests/metal_device_integration_tests.rs"
 
 [[example]]
 name = "gpu_validation"
