+++ crates/bitnet-opencl-work-size/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "bitnet-opencl-work-size"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "OpenCL work-size optimization heuristics for BitNet kernels"
+
+[dependencies]
+
